Makes Shifts a module
