+++ lib/shifts.rb
@@ -1,24 +1,14 @@
 require "./lib/keys"
 require "./lib/offsets"
 
-class Shifts
+module Shifts
   include Keys
   include Offsets
-  attr_reader :keys,
-              :offsets,
-              :shifts
-  def initialize
-    @keys = {a: nil,
-            b: nil,
-            c: nil,
-            d: nil}
-    @offsets = {a: nil,
-            b: nil,
-            c: nil,
-            d: nil}
-    @shifts = {a: nil,
-            b: nil,
-            c: nil,
-            d: nil}
-  end
+
+    def create_shifts
+      @shifts[:a] = @keys[:a] + @offsets[:a]
+      @shifts[:b] = @keys[:b] + @offsets[:b]
+      @shifts[:c] = @keys[:c] + @offsets[:c]
+      @shifts[:d] = @keys[:d] + @offsets[:d]
+    end
 end
